Move TaskStub to a nested class in TaskListTest
Putting TaskStub on the top level clutters the package and makes
it harder to identify test classes.

Let's move it to a nested class, putting it inside TaskListTest.
This makes sense because TaskListTest directly requires a stub for task,
and doing this clears up the package.
