+++ src/test/java/duke/logic/tasks/TaskListTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import org.junit.jupiter.api.Test;
 
+import duke.logic.commands.UpdateCommand.UpdateTaskDescriptor;
 import duke.logic.commands.UpdateTaskDescriptorBuilder;
 
 public class TaskListTest {
@@ -80,4 +81,43 @@ public class TaskListTest {
         s.append("Total: ").append(4).append(" tasks");
         assertEquals(s.toString(), list.toString());
     }
+
+    /**
+     * A stub for Task.
+     */
+    public static class TaskStub extends Task {
+        private final int stubNo;
+
+        public TaskStub(int stubNo) {
+            super("Stub task " + stubNo, true);
+            this.stubNo = stubNo;
+        }
+
+        @Override
+        public Task createUpdatedCopy(UpdateTaskDescriptor updateDescriptor) {
+            return new TaskStub(stubNo + 10);
+        }
+
+        @Override
+        public String getStatusIcon() {
+            return "X";
+        }
+
+        @Override
+        public String getSaveFormat() {
+            return "S | 1 | Stub task " + stubNo;
+        }
+
+        @Override
+        public String toString() {
+            return "[S][X] Stub task " + stubNo;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            return obj == this
+                    || (obj instanceof TaskStub && stubNo == ((TaskStub) obj).stubNo);
+        }
+    }
+
 }
